// src/types/team.types.ts
export type TeamHero = {
  eyebrow: string
  title: string
  description: string
  imageUrl: string
}

export type Leader = {
  id: string
  name: string
  position: string
  imageUrl: string
  careers: string[]
}

export type TeamMember = {
  id: string
  name: string
  position: string
  imageUrl: string
  careers: string[]
}